backend/pyproject.toml
[project]
name = "sounds-good-backend"
version = "0.1.0"
description = "Sounds Good"
authors = [
    {name = "Arda Arslan",email = "arda.arslan@std.bogazici.edu.tr"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "groq (>=0.11.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "ruff (>=0.15.6,<0.16.0)",
    "black (>=26.3.1,<27.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"